SqlBuilder extracted and working
