+++ index.js
@@ -1,6 +1,7 @@
 const DataBroker = require('./src/DataBroker/DataBroker.js');
 const Db = require('./src/Db/Db.js');
+const SqlBuilder = require('./src/SqlBuilder/SqlBuilder.js');
 const Parser = require('./src/Parser/Parser.js');
 const Select = require('./src/Select/Select.js');
 
-module.exports = { DataBroker, Db, Parser, Select };
+module.exports = { DataBroker, Db, SqlBuilder, Parser, Select };
