+++ src/Db/Db.spec.js
@@ -832,76 +832,6 @@ describe('Db', () => {
 			);
 		});
 	});
-	describe('buildWhere()', () => {
-		it('should handle single strings', async () => {
-			const sql = db.buildWhere('status = 5');
-			expect(sql).toBe('status = 5');
-		});
-		it('should handle single values', async () => {
-			const sql = db.buildWhere('status', 5);
-			expect(sql).toBe('`status` = 5');
-		});
-		it('should handle single strings with expressions', async () => {
-			const sql = db.buildWhere('SUM(size) < 1024');
-			expect(sql).toBe('SUM(size) < 1024');
-		});
-		it('should handle between', async () => {
-			const sql = db.buildWhere('status BETWEEN', [1, 3]);
-			expect(sql).toBe('`status` BETWEEN 1 AND 3');
-		});
-		it('should handle expressions', async () => {
-			const sql = db.buildWhere('COUNT(*) BETWEEN', [1, 3]);
-			expect(sql).toBe('COUNT(*) BETWEEN 1 AND 3');
-		});
-		it('should handle >', async () => {
-			const sql = db.buildWhere('id >', 5);
-			expect(sql).toBe('`id` > 5');
-		});
-		it('should handle !=', async () => {
-			const sql = db.buildWhere('id !=', 5);
-			expect(sql).toBe('`id` != 5');
-		});
-		it('should handle null', async () => {
-			const sql = db.buildWhere('deleted_at', null);
-			expect(sql).toBe('`deleted_at` IS NULL');
-		});
-		it('should handle implicit IN()', async () => {
-			const sql = db.buildWhere('status', [1, 3]);
-			expect(sql).toBe('`status` IN(1,3)');
-		});
-		it('should handle explicit IN()', async () => {
-			const sql = db.buildWhere('status IN', [1, 3]);
-			expect(sql).toBe('`status` IN(1,3)');
-		});
-		it('should handle implicit NOT IN()', async () => {
-			const sql = db.buildWhere('status !=', [1, 3]);
-			expect(sql).toBe('`status` NOT IN(1,3)');
-		});
-		it('should handle explicit NOT IN()', async () => {
-			const sql = db.buildWhere('status NOT IN', [1, 3]);
-			expect(sql).toBe('`status` NOT IN(1,3)');
-		});
-		it('should handle IS NULL', async () => {
-			const sql = db.buildWhere('status', null);
-			expect(sql).toBe('`status` IS NULL');
-		});
-		it('should handle equals IS NULL', async () => {
-			const sql = db.buildWhere('status =', null);
-			expect(sql).toBe('`status` IS NULL');
-		});
-		it('should handle IS NOT NULL', async () => {
-			const sql = db.buildWhere('status !=', null);
-			expect(sql).toBe('`status` IS NOT NULL');
-		});
-		it('should check hasOwnProperty', async () => {
-			const obj = Object.create({
-				a: 1,
-			});
-			obj.b = 2;
-			const sql = db.buildWheres(obj);
-			expect(sql).toBe('`b` = 2');
-		});
-	});
 	describe('tpl() escaping', () => {
 		it('should template Numbers', async () => {
 			const { select } = db.tpl();
